requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-unfold==0.68.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
Faker==37.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
